feat(): reactor extracting LlmProviderPanel first pass
