+++ src/main/java/com/devoxx/genie/ui/DevoxxGenieToolWindowContent.java
@@ -1,31 +1,21 @@
 package com.devoxx.genie.ui;
 
-import com.devoxx.genie.chatmodel.ChatModelFactory;
-import com.devoxx.genie.chatmodel.ChatModelFactoryProvider;
 import com.devoxx.genie.model.Constant;
 import com.devoxx.genie.model.LanguageModel;
 import com.devoxx.genie.model.enumarations.ModelProvider;
 import com.devoxx.genie.model.request.ChatMessageContext;
 import com.devoxx.genie.service.*;
-import com.devoxx.genie.ui.component.JHoverButton;
 import com.devoxx.genie.ui.component.PromptInputArea;
 import com.devoxx.genie.ui.listener.ConversationEventListener;
 import com.devoxx.genie.ui.listener.CustomPromptChangeListener;
-import com.devoxx.genie.ui.listener.LLMSettingsChangeListener;
 import com.devoxx.genie.ui.listener.SettingsChangeListener;
-import com.devoxx.genie.ui.panel.ActionButtonsPanel;
-import com.devoxx.genie.ui.panel.ConversationPanel;
-import com.devoxx.genie.ui.panel.PromptContextFileListPanel;
-import com.devoxx.genie.ui.panel.PromptOutputPanel;
+import com.devoxx.genie.ui.panel.*;
 import com.devoxx.genie.ui.renderer.ModelInfoRenderer;
 import com.devoxx.genie.ui.settings.DevoxxGenieStateService;
 import com.devoxx.genie.ui.topic.AppTopics;
-import com.devoxx.genie.ui.util.NotificationUtil;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.ComboBox;
-import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.ui.Splitter;
 import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.wm.ToolWindow;
@@ -39,17 +29,14 @@ import org.jetbrains.annotations.NotNull;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
-import java.util.List;
 import java.util.*;
 
 import static com.devoxx.genie.model.Constant.MESSAGES;
-import static com.devoxx.genie.ui.util.DevoxxGenieIconsUtil.RefreshIcon;
 
 /**
  * The Devoxx Genie Tool Window Content.
  */
 public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
-                                                     LLMSettingsChangeListener,
                                                      ConversationStarter,
                                                      CustomPromptChangeListener,
                                                      ConversationEventListener {
@@ -64,9 +51,8 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
 
     @Getter
     private final JPanel contentPanel = new JPanel();
-    private final ComboBox<ModelProvider> modelProviderComboBox = new ComboBox<>();
-    private final ComboBox<LanguageModel> modelNameComboBox = new ComboBox<>();
 
+    private LlmProviderPanel llmProviderPanel;
     private ConversationPanel conversationPanel;
     private PromptInputArea promptInputArea;
     private PromptOutputPanel promptOutputPanel;
@@ -74,14 +60,9 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
     private ActionButtonsPanel actionButtonsPanel;
 
     private boolean isInitializationComplete = false;
-    private boolean isUpdatingModelNames = false;
 
-    private String lastSelectedProvider = null;
-    private String lastSelectedLanguageModel = null;
     private final ChatService chatService;
     private final ConversationStorageService storageService = ConversationStorageService.getInstance();
-
-    private final JButton refreshButton = new JHoverButton(RefreshIcon, true);
 
     /**
      * The Devoxx Genie Tool Window Content constructor.
@@ -103,46 +84,8 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
     private void onStateLoaded() {
 
         if (!isInitializationComplete) {
-            lastSelectedProvider = DevoxxGenieStateService.getInstance().getSelectedProvider(project.getLocationHash());
-            lastSelectedLanguageModel = DevoxxGenieStateService.getInstance().getSelectedLanguageModel(project.getLocationHash());
-
             setupUI();
-            restoreLastSelectedProvider();
-            restoreLastSelectedLanguageModel();
             isInitializationComplete = true;
-        }
-    }
-
-    /**
-     * Restore the last selected provider from persistent storage
-     */
-    private void restoreLastSelectedProvider() {
-        if (lastSelectedProvider != null) {
-            for (int i = 0; i < modelProviderComboBox.getItemCount(); i++) {
-                ModelProvider provider = modelProviderComboBox.getItemAt(i);
-                if (provider.getName().equals(lastSelectedProvider)) {
-                    modelProviderComboBox.setSelectedIndex(i);
-                    updateModelNamesComboBox(lastSelectedProvider);
-                    break;
-                }
-            }
-        } else {
-            setLastSelectedProvider();
-        }
-    }
-
-    /**
-     * Restore the last selected language model from persistent storage
-     */
-    private void restoreLastSelectedLanguageModel() {
-        if (lastSelectedLanguageModel != null) {
-            for (int i = 0; i < modelNameComboBox.getItemCount(); i++) {
-                LanguageModel model = modelNameComboBox.getItemAt(i);
-                if (model.getModelName().equals(lastSelectedLanguageModel)) {
-                    modelNameComboBox.setSelectedIndex(i);
-                    break;
-                }
-            }
         }
     }
 
@@ -153,21 +96,10 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
      */
     private void setupMessageBusConnection(@NotNull ToolWindow toolWindow) {
         MessageBusConnection messageBusConnection = project.getMessageBus().connect();
-        messageBusConnection.subscribe(AppTopics.LLM_SETTINGS_CHANGED_TOPIC, this);
+        messageBusConnection.subscribe(AppTopics.LLM_SETTINGS_CHANGED_TOPIC, this.llmProviderPanel);
         messageBusConnection.subscribe(AppTopics.CUSTOM_PROMPT_CHANGED_TOPIC, this);
         messageBusConnection.subscribe(AppTopics.CONVERSATION_TOPIC, this);
         Disposer.register(toolWindow.getDisposable(), messageBusConnection);
-    }
-
-    /**
-     * Set the last selected LLM provider or show default.
-     */
-    private void setLastSelectedProvider() {
-        ModelProvider modelProvider = modelProviderComboBox.getItemAt(0);
-        if (modelProvider != null) {
-            DevoxxGenieStateService.getInstance().setSelectedProvider(project.getLocationHash(), modelProvider.getName());
-            updateModelNamesComboBox(modelProvider.getName());
-        }
     }
 
     /**
@@ -180,7 +112,8 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
     }
 
     private void initializeComponents() {
-        modelNameComboBox.setRenderer(new ModelInfoRenderer());
+        llmProviderPanel = new LlmProviderPanel(project);
+        llmProviderPanel.getModelNameComboBox().setRenderer(new ModelInfoRenderer());
         promptInputArea = new PromptInputArea(resourceBundle, project);
         promptOutputPanel = new PromptOutputPanel(resourceBundle);
         promptContextFileListPanel = new PromptContextFileListPanel(project);
@@ -194,8 +127,7 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
     }
 
     private void setupListeners() {
-        modelNameComboBox.addActionListener(this::processModelNameSelection);
-        modelProviderComboBox.addActionListener(this::handleModelProviderSelectionChange);
+        llmProviderPanel.getModelNameComboBox().addActionListener(this::processModelNameSelection);
     }
 
     /**
@@ -205,7 +137,7 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
      */
     private @NotNull JPanel createTopPanel() {
         JPanel topPanel = new JPanel(new BorderLayout());
-        topPanel.add(createSelectionPanel(), BorderLayout.NORTH);
+        topPanel.add(llmProviderPanel, BorderLayout.NORTH);
         topPanel.add(conversationPanel, BorderLayout.CENTER);
         return topPanel;
     }
@@ -228,110 +160,25 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
      */
     @Override
     public void settingsChanged(boolean hasKey) {
-        ModelProvider currentProvider = (ModelProvider) modelProviderComboBox.getSelectedItem();
-        LanguageModel currentModel = (LanguageModel) modelNameComboBox.getSelectedItem();
+        ModelProvider currentProvider = (ModelProvider) llmProviderPanel.getModelProviderComboBox().getSelectedItem();
+        LanguageModel currentModel = (LanguageModel) llmProviderPanel.getModelNameComboBox().getSelectedItem();
 
-        modelProviderComboBox.removeAllItems();
-        modelNameComboBox.removeAllItems();
-        addModelProvidersToComboBox();
+        llmProviderPanel.getModelProviderComboBox().removeAllItems();
+        llmProviderPanel.getModelNameComboBox().removeAllItems();
+        llmProviderPanel.addModelProvidersToComboBox();
 
         if (currentProvider != null) {
-            modelProviderComboBox.setSelectedItem(currentProvider);
-            updateModelNamesComboBox(currentProvider.getName());
+            llmProviderPanel.getModelProviderComboBox().setSelectedItem(currentProvider);
+            llmProviderPanel.updateModelNamesComboBox(currentProvider.getName());
 
             if (currentModel != null) {
-                modelNameComboBox.setSelectedItem(currentModel);
+                llmProviderPanel.getModelNameComboBox().setSelectedItem(currentModel);
             }
         } else {
-            setLastSelectedProvider();
+            llmProviderPanel.setLastSelectedProvider();
         }
 
         actionButtonsPanel.configureSearchButtonsVisibility();
-    }
-
-    /**
-     * Add the LLM providers to combobox.
-     * Only show the cloud-based LLM providers for which we have an API Key.
-     */
-    private void addModelProvidersToComboBox() {
-        LLMProviderService providerService = LLMProviderService.getInstance();
-        providerService.getAvailableModelProviders().stream()
-                .distinct()
-                .sorted(Comparator.comparing(ModelProvider::getName))
-                .forEach(modelProviderComboBox::addItem);
-    }
-
-    /**
-     * Create the LLM and model name selection panel.
-     *
-     * @return the selection panel
-     */
-    private @NotNull JPanel createSelectionPanel() {
-        JPanel toolPanel = new JPanel();
-        toolPanel.setLayout(new BoxLayout(toolPanel, BoxLayout.Y_AXIS));
-
-        addModelProvidersToComboBox();
-
-        JPanel providerPanel = createProviderPanel();
-        toolPanel.add(providerPanel);
-
-        toolPanel.add(Box.createVerticalStrut(5));
-
-        JPanel modelPanel = new JPanel(new BorderLayout());
-        modelPanel.add(modelNameComboBox, BorderLayout.CENTER);
-        toolPanel.add(modelPanel);
-
-        // Set preferred and maximum sizes for both combo boxes
-        Dimension comboBoxSize = new Dimension(Integer.MAX_VALUE, modelProviderComboBox.getPreferredSize().height);
-        modelProviderComboBox.setMaximumSize(comboBoxSize);
-        modelNameComboBox.setMaximumSize(comboBoxSize);
-
-        return toolPanel;
-    }
-
-    /**
-     * Create the LLM provider panel.
-     *
-     * @return the provider panel
-     */
-    private @NotNull JPanel createProviderPanel() {
-        JPanel providerPanel = new JPanel(new BorderLayout());
-        providerPanel.add(modelProviderComboBox, BorderLayout.CENTER);
-
-        refreshButton.setToolTipText("Refresh models");
-        refreshButton.addActionListener(e -> refreshModels());
-
-        providerPanel.add(refreshButton, BorderLayout.EAST);
-        return providerPanel;
-    }
-
-    /**
-     * Refresh the list of local models
-     */
-    private void refreshModels() {
-        ModelProvider selectedProvider = (ModelProvider) modelProviderComboBox.getSelectedItem();
-        if (selectedProvider == null) {
-            return;
-        }
-
-        if (selectedProvider == ModelProvider.LMStudio || selectedProvider == ModelProvider.Ollama || selectedProvider == ModelProvider.Jan) {
-            ApplicationManager.getApplication().invokeLater(() -> {
-                refreshButton.setEnabled(false);
-
-                ChatModelFactory factory = ChatModelFactoryProvider.getFactoryByProvider(selectedProvider.name())
-                        .orElseThrow(() -> new IllegalArgumentException("No factory for provider: " + selectedProvider));
-                factory.resetModels();
-
-                updateModelNamesComboBox(selectedProvider.getName());
-                modelNameComboBox.setRenderer(new ModelInfoRenderer());
-                modelNameComboBox.revalidate();
-                modelNameComboBox.repaint();
-                refreshButton.setEnabled(true);
-
-            });
-        } else {
-            NotificationUtil.sendNotification(project, "Model refresh is only available for LMStudio, Ollama and Jan providers.");
-        }
     }
 
     /**
@@ -359,8 +206,8 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
         actionButtonsPanel = new ActionButtonsPanel(project,
             promptInputArea,
             promptOutputPanel,
-            modelProviderComboBox,
-            modelNameComboBox,
+                llmProviderPanel.getModelProviderComboBox(),
+                llmProviderPanel.getModelNameComboBox(),
             this);
         return actionButtonsPanel;
     }
@@ -393,97 +240,12 @@ public class DevoxxGenieToolWindowContent implements SettingsChangeListener,
     private void processModelNameSelection(@NotNull ActionEvent e) {
 
         if (e.getActionCommand().equals(Constant.COMBO_BOX_CHANGED) && isInitializationComplete) {
-            LanguageModel selectedModel = (LanguageModel) modelNameComboBox.getSelectedItem();
+            LanguageModel selectedModel = (LanguageModel) llmProviderPanel.getModelNameComboBox().getSelectedItem();
             if (selectedModel != null) {
                 DevoxxGenieStateService.getInstance().setSelectedLanguageModel(project.getLocationHash(), selectedModel.getModelName());
                 actionButtonsPanel.updateTokenUsage(selectedModel.getContextWindow());
             }
         }
-    }
-
-    /**
-     * Process the model provider selection change.
-     * Set the model provider and update the model names.
-     */
-    private void handleModelProviderSelectionChange(@NotNull ActionEvent e) {
-        if (!e.getActionCommand().equals(Constant.COMBO_BOX_CHANGED) || !isInitializationComplete || isUpdatingModelNames)
-            return;
-
-        isUpdatingModelNames = true;
-
-        try {
-            JComboBox<?> comboBox = (JComboBox<?>) e.getSource();
-            ModelProvider modelProvider = (ModelProvider) comboBox.getSelectedItem();
-            if (modelProvider != null) {
-                DevoxxGenieStateService.getInstance().setSelectedProvider(project.getLocationHash(), modelProvider.getName());
-
-                updateModelNamesComboBox(modelProvider.getName());
-                modelNameComboBox.setRenderer(new ModelInfoRenderer());
-                modelNameComboBox.revalidate();
-                modelNameComboBox.repaint();
-            }
-        } finally {
-            isUpdatingModelNames = false;
-        }
-    }
-
-    /**
-     * Update the model names combobox.
-     */
-    private void updateModelNamesComboBox(String modelProvider) {
-        Optional.ofNullable(modelProvider).ifPresent(provider -> {
-            try {
-                modelNameComboBox.removeAllItems();
-                modelNameComboBox.setVisible(true);
-
-                ChatModelFactoryProvider
-                    .getFactoryByProvider(provider)
-                    .ifPresentOrElse(
-                        factory -> {
-                            List<LanguageModel> models = factory.getModels();
-                            if (models.isEmpty()) {
-                                hideModelNameComboBox();
-                            } else {
-                                populateModelNames(factory);
-                            }
-                        },
-                        this::hideModelNameComboBox
-                    );
-            } catch (Exception e) {
-                LOG.error("Error updating model names", e);
-                Messages.showErrorDialog(project, "Failed to update model names: " + e.getMessage(), "Error");
-            }
-        });
-    }
-
-    /**
-     * Populate the model names.
-     *
-     * @param chatModelFactory the chat model factory
-     */
-    private void populateModelNames(@NotNull ChatModelFactory chatModelFactory) {
-        modelNameComboBox.removeAllItems();
-        List<LanguageModel> modelNames = new ArrayList<>(chatModelFactory.getModels());
-        if (modelNames.isEmpty()) {
-            hideModelNameComboBox();
-        } else {
-            modelNames.sort(Comparator.naturalOrder());
-            modelNames.forEach(modelNameComboBox::addItem);
-        }
-    }
-
-    /**
-     * Hide the model name combobox.
-     */
-    private void hideModelNameComboBox() {
-        modelNameComboBox.setVisible(false);
-    }
-
-    @Override
-    public void settingsChanged() {
-        updateModelNamesComboBox(
-            DevoxxGenieStateService.getInstance().getSelectedProvider(project.getLocationHash())
-        );
     }
 
     @Override
