+++ src/main/java/com/devoxx/genie/ui/listener/LLMSettingsChangeListener.java
@@ -1,5 +1,5 @@
 package com.devoxx.genie.ui.listener;
 
 public interface LLMSettingsChangeListener {
-    void settingsChanged();
+    void llmSettingsChanged();
 }
